Feat form(Form UI):[MCS-Form] - add validation

+++ netomichatsdk/src/main/java/com/netomi/chat/model/messages/NCWGenericChannelResponse.kt
@@ -32,7 +32,7 @@ data class NCWAttachmentListRequest(
 
     var type: String? = null,
     var title: String? = null,
-    var timestamp: Int? = null,
+    var timestamp: Long? = null,
     var largeImageUrl: String? = null,
     var isReviewEnabled: Boolean? = null,
     var thumbnailUrl: String? = null,
